senselet++Onewire_Server/requirements.txt
influxdb==5.3.1
paho_mqtt==1.6.1
pytz==2021.3
requests==2.27.1
